[refactor]:将所有表示文件长度的long都修改为ulong

+++ Assets/CatAsset/Editor/BuildInfo/AssetBuildInfo.cs
@@ -25,12 +25,12 @@ namespace CatAsset.Editor
         /// <summary>
         /// 资源文件长度
         /// </summary>
-        public long Length;
+        public ulong Length;
         
         public AssetBuildInfo(string name)
         {
             Name = name;
-            Length = new FileInfo(name).Length;
+            Length = (ulong)new FileInfo(name).Length;
         }
 
         public override string ToString()
